my personal

+++ app/Http/Controllers/admin/UserController.php
@@ -5,12 +5,16 @@ namespace App\Http\Controllers\admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\User;
+use App\Models\City;
+use App\Models\State;
+use App\Models\Payment;
 use App\Models\Role as UserRole;
 use Spatie\Permission\Models\Role;
 use DB;
 use Auth;
 use Hash;
 use Illuminate\Support\Arr;
+use Session;
 
 class UserController extends Controller
 {
@@ -29,7 +33,7 @@ class UserController extends Controller
     public function index(Request $request)
     {
         if($request->ajax()){
-            $query = User::orderby('id', 'desc')->where('id', '>', 0);
+            $query = User::orderby('id', 'asc')->where('id', '>', 0);
             if($request['search'] != ""){
                 $query->where('name', 'like', '%'. $request['search'] .'%')
                     ->orWhere('last_name', 'like', '%'. $request['search'].'%')
@@ -42,11 +46,11 @@ class UserController extends Controller
                 }
                 $query->where('status', $request['status']);
             }
-            $users = $query->paginate(5);
+            $users = $query->paginate(10);
             return (string) view('admin.user.search', compact('users'));
         }
         $page_title = 'All Users';
-        $users = User::orderBy('id','DESC')->paginate(5);
+        $users = User::orderBy('id','asc')->paginate(10);
         return view('admin.user.index', compact('users','page_title'));
     }
 
@@ -58,7 +62,7 @@ class UserController extends Controller
     public function create()
     {
         $page_title = 'Add User';
-        $roles = Role::orderby('id', 'desc')->get(['name', 'id']);
+        $roles = Role::orderby('id', 'asc')->get(['name', 'id']);
         return view('admin.user.create',compact('roles','page_title'));
     }
 
@@ -83,8 +87,7 @@ class UserController extends Controller
         $user = User::create($input);
         $user->assignRole($request->input('roles'));
 
-        return redirect()->route('user.index')
-                        ->with('message','User created successfully');
+        return redirect()->route('user.index')->with('message','User created successfully');
     }
 
     /**
@@ -111,6 +114,7 @@ class UserController extends Controller
         $user = User::with('roles')->find($id);
         $roles = Role::orderby('id', 'desc')->get(['name', 'id']);
         $userRole = $user->roles->pluck('name','name')->all();
+        $user =  User::where('id', $id)->first();
         return view('admin.user.edit',compact('user','roles','userRole', 'page_title'));
     }
 
@@ -123,28 +127,41 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'name' => 'required|max:200',
-            'email' => 'required|max:200|email|unique:users,email,'.$id,
-            'password' => 'same:confirm-password',
-            'roles' => 'required'
-        ]);
+        $user = User::where('id', $id)->first();
+        if($id){
+            $this->validate($request, [
+                'status' => 'required',
+                'top_rated' => 'required',
 
-        $input = $request->all();
-        if(!empty($input['password'])){
-            $input['password'] = Hash::make($input['password']);
-        }else{
-            $input = Arr::except($input,array('password'));
+            ]);
+
+            $user->status = $request->status;
+            $user->top_rated = $request->top_rated;
+            $user->update();
+        }
+        else{
+           $this->validate($request, [
+                'name' => 'required|max:200',
+                'email' => 'required|max:200|email|unique:users,email,'.$id,
+                'password' => 'same:confirm-password',
+                'roles' => 'required'
+            ]);
+
+            $input = $request->all();
+            if(!empty($input['password'])){
+                $input['password'] = Hash::make($input['password']);
+            }else{
+                $input = Arr::except($input,array('password'));
+            }
+
+            $user = User::find($id);
+            $user->update($input);
+            DB::table('model_has_roles')->where('model_id', $id)->delete();
+
+            $user->assignRole($request->input('roles'));
         }
 
-        $user = User::find($id);
-        $user->update($input);
-        DB::table('model_has_roles')->where('model_id', $id)->delete();
-
-        $user->assignRole($request->input('roles'));
-
-        return redirect()->route('user.index')
-                        ->with('message','User updated successfully');
+        return redirect()->route('user.index')->with('message','User updated successfully');
     }
 
     /**
@@ -162,23 +179,46 @@ class UserController extends Controller
     }
 
     public function userEditProfile()
-    {
-        return view('website.user-dashboard.edit');
+    {   $cities = City::where('status', 1)->get();
+        $states = State::where('city_id')->get();
+        $user =  User::where('id', Auth::user()->id)->first();
+        return view('website.user-dashboard.edit', compact('cities', 'states','user'));
     }
 
     public function userUpdateProfile(Request $request)
     {
         $user = User::where('id', Auth::user()->id)->first();
         $user->name = $request->first_name;
-      
+        $user->middle_name = $request->middle_name;
+        $user->last_name = $request->last_name;
+        $user->address = $request->address;
+        $user->designation = $request->designation;
+        $user->about_me = $request->about_me;
+        $user->whatsapp = $request->whatsapp;
+        $user->skype = $request->skype;
+        $user->facebook = $request->facebook;
+        $user->twitter = $request->twitter;
+        $user->instagram = $request->instagram;
+        $user->behance = $request->behance;
+        $user->youtube = $request->youtube;
+        $user->city_id = $request->city_id;
+        $user->state_id = $request->state_id;
+        $user->zip_code = $request->zip_code;
+        $user->license = $request->license;
+        if(isset($request->image)){
+            $photo= date('YmdHis').'.'.$request->file('image')->getClientOriginalExtension();
+            $request->image->move(public_path('/admin/assets/images/UserImage') , $photo);
+            $user->image = $photo;
+        }
+
         if(empty($request->name)){
             $this->validate($request, [
                 'first_name' => 'required',
+                'city_id' => 'required',
+                'state_id' => 'required',
             ]);
         }
-        
-        $user->last_name = $request->last_name;
-       
+
         if(isset($request->password)){
             $this->validate($request, [
                 'first_name' => 'required',
@@ -192,9 +232,29 @@ class UserController extends Controller
         return redirect()->back()->with('message','Profile updated successfully');
     }
 
+    public function authenticate(Request $request)
+    {
+        $user = User::where('email', $request->email)->first();
+
+        if(!empty($user) && $user->hasRole($request->user_type)){
+            $credentials = $request->only('email', 'password');
+            if (Auth::attempt($credentials)) {
+                $user_id = Auth::user()->id;
+                $isuser_paid = Payment::where('customer_id',$user_id)->where('payment_status','succeeded')->first();
+                if(empty($isuser_paid)){
+                    return redirect()->route('stripe');
+                }else{
+                    return redirect()->route('dashboard');
+                }
+            }
+            return redirect()->back()->with('error', 'Failed to login try again.!');
+        }else{
+            return redirect()->back()->with('error', 'Something went wrong!');
+        }
+    }
     public function logOut()
     {
         Auth::logout();
-        return redirect()->route('login');
+        return redirect()->route('admin.login');
     }
 }
